+++ src/Module.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\debug;
@@ -60,7 +60,7 @@ class Module extends \yii\base\Module implements BootstrapInterface
      */
     public $controllerNamespace = 'yii\debug\controllers';
     /**
-     * @var LogTarget
+     * @var LogTarget|array|string the logTarget object, or the configuration for creating the logTarget object.
      */
     public $logTarget;
     
@@ -113,6 +113,12 @@ class Module extends \yii\base\Module implements BootstrapInterface
      * @since 2.1.1
      */
     public $defaultHeight = 50;
+    /**
+     * @var string toolbar position on web page. Use 'bottom' or 'upper'.
+     * You may add custom value via .yii-debug-toolbar_position_{yourValue} css.
+     * @since 2.1.14
+     */
+    public $toolbarPosition = 'bottom';
     /**
      * @var bool whether to enable message logging for the requests about debug module actions.
      * You normally do not want to keep these logs because they may distract you from the logs about your applications.
@@ -230,9 +236,7 @@ class Module extends \yii\base\Module implements BootstrapInterface
         parent::init();
         $this->dataPath = Yii::getAlias($this->dataPath);
 
-        if (Yii::$app instanceof \yii\web\Application) {
-            $this->initPanels();
-        }
+        $this->initPanels();
     }
 
     /**
@@ -272,8 +276,16 @@ class Module extends \yii\base\Module implements BootstrapInterface
      */
     public function bootstrap($app)
     {
+        if (is_array($this->logTarget)) {
+            if (!isset($this->logTarget['class'])) {
+                $this->logTarget['class'] = 'yii\debug\LogTarget';
+            }
+            $this->logTarget = Yii::createObject($this->logTarget, [$this]);
+        } elseif (is_string($this->logTarget)) {
+            $this->logTarget = Yii::createObject($this->logTarget, [$this]);
+        }
         /* @var $app \yii\base\Application */
-        $this->logTarget = $app->getLog()->targets['debug'] = new LogTarget($this);
+        $app->getLog()->targets['debug'] = $this->logTarget;
 
         // delay attaching event handler to the view component after it is fully configured
         $app->on(Application::EVENT_BEFORE_REQUEST, function () use ($app) {
@@ -286,15 +298,15 @@ class Module extends \yii\base\Module implements BootstrapInterface
         $app->getUrlManager()->addRules([
             [
                 'class' => $this->urlRuleClass,
-                'route' => $this->id,
-                'pattern' => $this->id,
+                'route' => $this->getUniqueId(),
+                'pattern' => $this->getUniqueId(),
                 'normalizer' => false,
                 'suffix' => false
             ],
             [
                 'class' => $this->urlRuleClass,
-                'route' => $this->id . '/<controller>/<action>',
-                'pattern' => $this->id . '/<controller:[\w\-]+>/<action:[\w\-]+>',
+                'route' => $this->getUniqueId() . '/<controller>/<action>',
+                'pattern' => $this->getUniqueId() . '/<controller:[\w\-]+>/<action:[\w\-]+>',
                 'normalizer' => false,
                 'suffix' => false
             ]
@@ -348,7 +360,7 @@ class Module extends \yii\base\Module implements BootstrapInterface
             return;
         }
         $url = Url::toRoute([
-            '/' . $this->id . '/default/view',
+            '/' . $this->getUniqueId() . '/default/view',
             'tag' => $this->logTarget->tag,
         ]);
         $event->sender->getHeaders()
@@ -372,7 +384,7 @@ class Module extends \yii\base\Module implements BootstrapInterface
     public function getToolbarHtml()
     {
         $url = Url::toRoute([
-            '/' . $this->id . '/default/toolbar',
+            '/' . $this->getUniqueId() . '/default/toolbar',
             'tag' => $this->logTarget->tag,
         ]);
 
@@ -398,7 +410,7 @@ class Module extends \yii\base\Module implements BootstrapInterface
 
         /* @var $view View */
         $view = $event->sender;
-        echo $view->renderDynamic('return Yii::$app->getModule("' . $this->id . '")->getToolbarHtml();');
+        echo $view->renderDynamic('return Yii::$app->getModule("' . $this->getUniqueId() . '")->getToolbarHtml();');
 
         // echo is used in order to support cases where asset manager is not available
         echo '<style>' . $view->renderPhpFile(__DIR__ . '/assets/css/toolbar.css') . '</style>';
@@ -465,20 +477,25 @@ class Module extends \yii\base\Module implements BootstrapInterface
      */
     protected function corePanels()
     {
-        return [
+        $corePanels = [
             'config' => ['class' => 'yii\debug\panels\ConfigPanel'],
-            'request' => ['class' => 'yii\debug\panels\RequestPanel'],
-            'router' => ['class' => 'yii\debug\panels\RouterPanel'],
             'log' => ['class' => 'yii\debug\panels\LogPanel'],
             'profiling' => ['class' => 'yii\debug\panels\ProfilingPanel'],
             'db' => ['class' => 'yii\debug\panels\DbPanel'],
             'event' => ['class' => 'yii\debug\panels\EventPanel'],
-            'assets' => ['class' => 'yii\debug\panels\AssetPanel'],
             'mail' => ['class' => 'yii\debug\panels\MailPanel'],
             'timeline' => ['class' => 'yii\debug\panels\TimelinePanel'],
-            'user' => ['class' => 'yii\debug\panels\UserPanel'],
             'dump' => ['class' => 'yii\debug\panels\DumpPanel'],
         ];
+
+        if (Yii::$app instanceof \yii\web\Application) {
+            $corePanels['router'] = ['class' => 'yii\debug\panels\RouterPanel'];
+            $corePanels['request'] = ['class' => 'yii\debug\panels\RequestPanel'];
+            $corePanels['user'] = ['class' => 'yii\debug\panels\UserPanel'];
+            $corePanels['asset'] = ['class' => 'yii\debug\panels\AssetPanel'];
+        }
+
+        return $corePanels;
     }
 
     /**