+++ src/models/timeline/DataProvider.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\debug\models\timeline;
@@ -13,7 +13,7 @@ use yii\debug\panels\TimelinePanel;
 /**
  * DataProvider implements a data provider based on a data array.
  *
- * @property-read array $rulers This property is read-only.
+ * @property-read array $rulers
  *
  * @author Dmitriy Bashkarev <dmitriy@bashkarev.com>
  * @since 2.0.8