+++ src/panels/TimelinePanel.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\debug\panels;
@@ -17,8 +17,8 @@ use yii\debug\Panel;
  * Debugger panel that collects and displays timeline data.
  *
  * @property array $colors
- * @property-read float $duration This property is read-only.
- * @property-read float $start This property is read-only.
+ * @property-read float $duration
+ * @property-read float $start
  * @property array $svgOptions
  *
  * @author Dmitriy Bashkarev <dmitriy@bashkarev.com>