+++ src/panels/UserPanel.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\debug\panels;
@@ -27,8 +27,8 @@ use yii\di\Instance;
 /**
  * Debugger panel that collects and displays user data.
  *
- * @property-read DataProviderInterface $userDataProvider This property is read-only.
- * @property-read Model|UserSearchInterface $usersFilterModel This property is read-only.
+ * @property-read DataProviderInterface $userDataProvider
+ * @property-read Model|UserSearchInterface $usersFilterModel
  *
  * @author Daniel Gomez Pan <pana_1990@hotmail.com>
  * @since 2.0.8
@@ -41,7 +41,7 @@ class UserPanel extends Panel
      * Settable: allow, roles, ips, matchCallback, denyCallback.
      * By default deny for everyone. Recommendation: can allow for administrator
      * or developer (if implement) role: ['allow' => true, 'roles' => ['admin']]
-     * @see http://www.yiiframework.com/doc-2.0/guide-security-authorization.html
+     * @see https://www.yiiframework.com/doc-2.0/guide-security-authorization.html
      * @since 2.0.10
      */
     public $ruleUserSwitch = [
@@ -59,7 +59,7 @@ class UserPanel extends Panel
     public $filterModel;
     /**
      * @var array allowed columns for GridView.
-     * @see http://www.yiiframework.com/doc-2.0/yii-grid-gridview.html#$columns-detail
+     * @see https://www.yiiframework.com/doc-2.0/yii-grid-gridview.html#$columns-detail
      * @since 2.0.10
      */
     public $filterColumns = [];
@@ -88,7 +88,7 @@ class UserPanel extends Panel
         $this->userSwitch = new UserSwitch(['userComponent' => $this->userComponent]);
         $this->addAccessRules();
 
-        if (!is_object($this->filterModel)
+        if (is_string($this->filterModel)
             && class_exists($this->filterModel)
             && in_array('yii\debug\models\search\UserSearchInterface', class_implements($this->filterModel), true)
         ) {
@@ -118,13 +118,13 @@ class UserPanel extends Panel
      */
     private function addAccessRules()
     {
-        $this->ruleUserSwitch['controllers'] = [$this->module->id . '/user'];
+        $this->ruleUserSwitch['controllers'] = [$this->module->getUniqueId() . '/user'];
 
         $this->module->attachBehavior(
             'access_debug',
             [
                 'class' => 'yii\filters\AccessControl',
-                'only' => [$this->module->id . '/user', $this->module->id . '/default'],
+                'only' => [$this->module->getUniqueId() . '/user', $this->module->getUniqueId() . '/default'],
                 'user' => $this->userSwitch->getMainUser(),
                 'rules' => [
                     $this->ruleUserSwitch,