+++ src/views/default/index.php
@@ -3,6 +3,7 @@
 use yii\data\ArrayDataProvider;
 use yii\grid\GridView;
 use yii\helpers\Html;
+use yii\helpers\Url;
 
 /* @var $this \yii\web\View */
 /* @var $manifest array */
@@ -16,7 +17,7 @@ $this->title = 'Yii Debugger';
     <div id="yii-debug-toolbar" class="yii-debug-toolbar yii-debug-toolbar_position_top" style="display: none;">
         <div class="yii-debug-toolbar__bar">
             <div class="yii-debug-toolbar__block yii-debug-toolbar__title">
-                <a href="#">
+                <a href="<?= Url::to(['index']) ?>">
                     <img width="30" height="30" alt="" src="<?= \yii\debug\Module::getYiiLogo() ?>">
                 </a>
             </div>
@@ -47,10 +48,6 @@ $this->title = 'Yii Debugger';
                 'filterModel' => $searchModel,
                 'rowOptions' => function ($model) use ($searchModel, $hasDbPanel) {
                     if ($searchModel->isCodeCritical($model['statusCode'])) {
-                        return ['class' => 'table-danger'];
-                    }
-
-                    if ($hasDbPanel && $this->context->module->panels['db']->isQueryCountCritical($model['sqlCount'])) {
                         return ['class' => 'table-danger'];
                     }
 
@@ -111,17 +108,26 @@ $this->title = 'Yii Debugger';
                             /* @var $dbPanel \yii\debug\panels\DbPanel */
                             $dbPanel = $this->context->module->panels['db'];
 
+                            $title = "Executed {$data['sqlCount']} database queries.";
+                            $warning = '';
                             if ($dbPanel->isQueryCountCritical($data['sqlCount'])) {
-                                $content = Html::tag('b', $data['sqlCount']) . ' ' . Html::tag('span', '&#x26a0;');
-
-                                return Html::a($content, ['view', 'panel' => 'db', 'tag' => $data['tag']], [
-                                    'title' => 'Too many queries. Allowed count is ' . $dbPanel->criticalQueryThreshold,
-                                ]);
-
+                                $warning .= 'Too many queries. Allowed count is ' . $dbPanel->criticalQueryThreshold;
                             }
-                            return $data['sqlCount'];
+                            if (!empty($data['excessiveCallersCount'])) {
+                                $warning .= ($warning ? ' &#10;' : '') . $data['excessiveCallersCount'] . ' '
+                                    . ($data['excessiveCallersCount'] == 1 ? 'caller is' : 'callers are')
+                                    . ' making too many calls.';
+                            }
+
+                            $content = $data['sqlCount'];
+                            if ($warning) {
+                                $content .= ' <span title="' . $warning . '">&#x26a0;</span>';
+                            }
+
+                            return '<a href="' . Url::to(['view', 'panel' => 'db', 'tag' => $data['tag']]) .'"
+                                        title="' . $title . '">' . $content . '</a>';
                         },
-                        'format' => 'html',
+                        'format' => 'raw',
                     ] : null,
                     [
                         'attribute' => 'mailCount',
@@ -134,7 +140,8 @@ $this->title = 'Yii Debugger';
                             'post' => 'POST',
                             'delete' => 'DELETE',
                             'put' => 'PUT',
-                            'head' => 'HEAD'
+                            'head' => 'HEAD',
+                            'command' => 'COMMAND'
                         ]
                     ],
                     [
@@ -146,16 +153,17 @@ $this->title = 'Yii Debugger';
                     ],
                     [
                         'attribute' => 'url',
-                        'label' => 'URL',
+                        'label' => 'URL/Command',
                     ],
                     [
                         'attribute' => 'statusCode',
                         'value' => function ($data) {
                             $statusCode = $data['statusCode'];
+                            $method = $data['method'];
                             if ($statusCode === null) {
                                 $statusCode = 200;
                             }
-                            if ($statusCode >= 200 && $statusCode < 300) {
+                            if (($statusCode >= 200 && $statusCode < 300) || ($method == 'COMMAND' && $statusCode == 0)) {
                                 $class = 'badge-success';
                             } elseif ($statusCode >= 300 && $statusCode < 400) {
                                 $class = 'badge-info';