+++ CHANGELOG.md
@@ -1,10 +1,62 @@
 Yii Framework 2 debug extension Change Log
 ==========================================
 
-2.1.19 under development
+2.1.26 under development
 ------------------------
 
-- Enh #459: Add `yii\debug\Module::$authManager` for config RBAC manager used by UserPanel to load current user roles and permissions, default is `authManager`. (jafaripur)
+- Enh #459: Add `yii\debug\Module::$authManager` for config RBAC manager used by UserPanel to load current user roles and permissions, default is `authManager` (jafaripur)
+- Enh #430: Allow to configure toolbar position via `Module::$toolbarPosition` property (sasha-x)
+- Bug #528: Fix `yii\debug\Panel::getTraceLine()` to handle backtrace for internal PHP functions (zymeli)
+
+
+2.1.25 September 26, 2023
+-------------------------
+
+- Bug #503: Fix accessing toolbar data if it's not available (xepozz)
+
+
+2.1.24 July 10, 2023
+--------------------
+
+- Bug #504: Reduced db panel warnings for "critical query threshold" and "excessive callers" (rhertogh)
+- Bug #506: Correctly handle null values for `DbPanel::$criticalQueryThreshold` and `::$excessiveCallerThreshold` (MarkoNV, rhertogh)
+- Bug #507: Convert Symfony mailer headers to string in Mail panel (squio)
+- Enh #512: Changed default value for `yii\debug\panels\DbPanel::$excessiveCallerThreshold` to `null` (rhertogh)
+
+
+2.1.23 May 22, 2023
+-------------------
+
+- Enh #498: Add yii2-symfonymailer Support in src/panels/MailPanel.php (vansari)
+- Enh #501: Show DB query call location in DB panel (rhertogh)
+
+
+2.1.22 November 18, 2022
+------------------------
+
+- Enh #493: Extend variable masking to `requestBody` (yus-ham)
+
+
+2.1.21 September 06, 2022
+-------------------------
+
+- Enh #490: Allow debugging of console apps (chicuongit913)
+
+
+2.1.20 September 04, 2022
+-------------------------
+
+- Bug #461: Do not crash on empty panel data (przepompownia)
+- Enh #484: Allow `logTarget` config to be set as either a string, an array, or an `LogTarget` object (Sarke)
+- Enh #486: Allow to configure debugger as a submodule (achretien)
+
+2.1.19 April 05, 2022
+---------------------
+
+- Bug #466: Remove opis\closure dependency for PHP 8.1 compatibility. Closures in logs still working (sartor)
+- Bug #470: Fix PHP 8.1 warnings about implicit conversion from float to int loses precision (mishamosher)
+- Enh #469: Add option to change default LogTarget (laxity7)
+- Enh #478: Add option to censor variables in the request panel (dbx12)
 
 
 2.1.18 August 09, 2021